move createDarCollectionMultipleUserProperties
