+++ src/test/java/org/broadinstitute/consent/http/db/DarCollectionDAOTest.java
@@ -16,7 +16,10 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.RandomUtils;
+import org.broadinstitute.consent.http.enumeration.ElectionStatus;
+import org.broadinstitute.consent.http.enumeration.ElectionType;
 import org.broadinstitute.consent.http.enumeration.OrganizationType;
+import org.broadinstitute.consent.http.enumeration.UserFields;
 import org.broadinstitute.consent.http.models.Consent;
 import org.broadinstitute.consent.http.models.Dac;
 import org.broadinstitute.consent.http.models.DarCollection;
@@ -515,6 +518,47 @@ public class DarCollectionDAOTest extends DAOTestHelper {
         new Date()
     );
     return institutionDAO.findInstitutionById(id);
+  }
+
+  private DarCollection createDarCollectionMultipleUserProperties() {
+    User user = createUser();
+    Integer userId = user.getUserId();
+    createUserProperty(userId, UserFields.SUGGESTED_SIGNING_OFFICIAL.getValue());
+    createUserProperty(userId, UserFields.SUGGESTED_INSTITUTION.getValue());
+    createUserProperty(userId, UserFields.ERA_STATUS.getValue());
+    String darCode = "DAR-" + RandomUtils.nextInt(100, 1000);
+    Integer collection_id = darCollectionDAO.insertDarCollection(darCode, user.getUserId(),
+        new Date());
+    Dataset dataset = createDataset();
+    DataAccessRequest dar = createDataAccessRequest(user.getUserId(), collection_id, darCode);
+    dataAccessRequestDAO.insertDARDatasetRelation(dar.getReferenceId(), dataset.getDataSetId());
+    Election cancelled = createCancelledAccessElection(dar.getReferenceId(),
+        dataset.getDataSetId());
+    Election access = createDataAccessElection(dar.getReferenceId(), dataset.getDataSetId());
+    createFinalVote(user.getUserId(), cancelled.getElectionId());
+    createFinalVote(user.getUserId(), access.getElectionId());
+    createDataAccessRequest(user.getUserId(), collection_id, darCode);
+    createDataAccessRequest(user.getUserId(), collection_id, darCode);
+    return darCollectionDAO.findDARCollectionByCollectionId(collection_id);
+  }
+
+  private void createUserProperty(Integer userId, String field) {
+    UserProperty property = new UserProperty();
+    property.setPropertyKey(field);
+    property.setPropertyValue(UUID.randomUUID().toString());
+    property.setUserId(userId);
+    userPropertyDAO.insertAll(List.of(property));
+  }
+
+  private Election createCancelledAccessElection(String referenceId, Integer datasetId) {
+    Integer electionId = electionDAO.insertElection(
+        ElectionType.DATA_ACCESS.getValue(),
+        ElectionStatus.CANCELED.getValue(),
+        new Date(),
+        referenceId,
+        datasetId
+    );
+    return electionDAO.findElectionById(electionId);
   }
 
 }
